software/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyserial>=3.5
joblib>=1.0.0
